Backend/authentication_service/requirements.txt
sqlalchemy
pydantic
fastapi
uvicorn
pymysql
passlib[bcrypt]
bcrypt==4.2.1
PyJWT
email-validator
